% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_lha_data.R
\name{create_lha_heatmap}
\alias{create_lha_heatmap}
\title{Create a local health area Geographic Map to visualize data}
\usage{
create_lha_heatmap(
  lha_data_frame,
  lha_id_column_name,
  colorby_column_name,
  title = "Data By Local Health Area",
  subtitle = colorby_column_name,
  colors = c("white", "yellow", "orange", "red", "firebrick"),
  color_breaks = c(0, 25, 50, 75, 100)
)
}
\arguments{
\item{lha_data_frame}{\itemize{
\item data.frame: the data for visualization in the map
}}

\item{lha_id_column_name}{\itemize{
\item character : the column name containing the LHA IDs
}}

\item{colorby_column_name}{\itemize{
\item character : the column name for initial visualization
}}

\item{title}{\itemize{
\item character : chart title (default = "Data By Local Health Area")
}}

\item{subtitle}{\itemize{
\item character : subtitle    (default = colorby_column_name value)
}}

\item{colors}{\itemize{
\item character vector: list of color values to use on the map.
can be specified as html names, hex codes, rgb or rgba string values
default = c("white", "yellow", "orange", "red", "firebrick")
}}

\item{color_breaks}{\itemize{
\item numeric vector: list of locations for the color breaks
must be the same length as the colors parameter
default = c(0, 25, 50, 75, 100)
}}
}
\value{
a CanvasXpress plot object
}
\description{
Create a local health area Geographic Map to visualize data
}
\examples{

\code{#TODO}

}
