% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-survey.R
\name{join_dictionary_with_survey_data}
\alias{join_dictionary_with_survey_data}
\title{Joining transposed survey data with the data dictionary}
\usage{
join_dictionary_with_survey_data(raw_survey_data, data_dictionary_file)
}
\arguments{
\item{transposed_survey_data}{}
}
\value{

}
\description{
Joining transposed survey data with the data dictionary
}
