% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-survey.R
\name{parse_survey_data_dictionary}
\alias{parse_survey_data_dictionary}
\title{Parse data dictionary with survey data}
\usage{
parse_survey_data_dictionary(joined_data_dict_transposed_tbl)
}
\arguments{
\item{joined_data_dict_transposed_tbl}{}
}
\value{

}
\description{
Parse data dictionary with survey data
}
