% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data-dict.R
\name{read_data_dictionary}
\alias{read_data_dictionary}
\title{Read in data dictionary sheet}
\usage{
read_data_dictionary(data_dictionary_file, sheet_name = "Data_dictionary")
}
\arguments{
\item{data_dictionary_file}{}

\item{sheet_name}{}
}
\value{
tibble
}
\description{
Read in data dictionary sheet
}
\examples{
data_dict_tbl <- read_data_dictionary_sheet("./data/abcd.xlsx")
}
