% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data-dict.R
\name{read_hsda_data_dict}
\alias{read_hsda_data_dict}
\title{Read HSDA information from data dictionary}
\usage{
read_hsda_data_dict(
  data_dictionary_file,
  sheet_name = "CHSA_LHA_HSDA_HA_lookup"
)
}
\arguments{
\item{data_dictionary_file}{}

\item{sheet_name}{}
}
\value{
tibble
}
\description{
Read HSDA information from data dictionary
}
