% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-survey.R
\name{transpose_survey_data}
\alias{transpose_survey_data}
\title{Transposing Survey Data}
\usage{
transpose_survey_data(raw_survey_data)
}
\arguments{
\item{raw_survey_data}{}
}
\value{

}
\description{
Transposing Survey Data
}
\examples{
Transposing Survey data table
Transposing the survey data to enable filtering the features 
substituting values from the dictionary
}
